Correctly parse the timestamp passed by grafana on 32bit systems
strconv.Atoi returns an int which is only guaranteed to be 32bit in size
and thus not big enough to store a grafana timestamp (ex: 1522838700000).
On 64bit systems int seems to be 64bit in size thus avoiding this issue.

strconv.ParseInt in contrast always returns an int64 and thus correctly
parses the timestamp on all systems.

Example:
https://play.golang.org/p/e3tg7Hb7Zy2

Reference:
https://golang.org/pkg/strconv/#Atoi
https://golang.org/pkg/strconv/#ParseInt
https://golang.org/pkg/builtin/#int