Add option to set report title in cmd mode.
This adds an additional command-line option -cmd-title which allows to
overwrite the default title (i.e. dashboard name). Depending on the dashboard
naming scheme, or the particular time-range of the report, the user might want
to set a custom title (like `Daily report for...') without having to use custom
templates.